fix some stuff in FlightParams.h

+++ src/FlightParams.h
@@ -3,6 +3,7 @@
 //TODO: CHECK EVERY PARAM! I want to see a nice green checkmark emoji when we confirm testing on a param! ✅
 
 
+#include <cstdint>
 constexpr float LAUNCH_THRESHHOLD_G = 5;
 constexpr float BURN_THRESHHOLD_G = 0.3;
 constexpr float IS_TUMBLING_VEL_THRESHOLD = 0.174; //10 dps in rad/s... idk the units for the gyro and this will obv need testing. Upper bound of hysterisis
@@ -22,10 +23,10 @@ constexpr uint32_t FLAP_EXTENDED_POS = 1500; //in PWM for fully extended flap
 constexpr uint32_t FLAP_RETRACTED_POS = 1000; //in PWM for fully retracted flap
 constexpr uint32_t AUGER_MAX_EXT_POS = 1500; //in PWM
 constexpr uint32_t AUGER_MAX_CLOSED_POS = 1000; //in PWM
-constexpr u_int32_t AUGER_EXT_POS_BAND =  50;
+constexpr uint32_t AUGER_EXT_POS_BAND =  50;
 constexpr uint32_t SOLID_DELIVERY_DOOR_OPEN_POS= 1500; //in PWM
 constexpr uint32_t SOLID_DELIVERY_DOOR_CLOSED_POS = 1000; //in PWM
 constexpr uint32_t SOLID_EJECTION_POS = 1500; //in PWM
 constexpr uint32_t LIQUID_DELIVERY_POS = 1500; //in PWM
 
-constexpr uint32_t DRILL_SPEED = 0.5; //im guessing m/s?
+constexpr float DRILL_SPEED = 0.5; //im guessing m/s?
